backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
google-generativeai==0.3.0
python-dotenv==1.0.0
httpx==0.25.0
pillow==10.1.0
numpy==1.24.3
